chore: 重构 desktop => home
