+++ src/app/home/Background/style.ts
