+++ src/app/home/Docker/AudioPlayer/Duration/style.ts
