+++ src/app/home/Docker/index.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import AudioPlayer from '@/app/desktop/Docker/AudioPlayer';
-import MessageInput from '@/app/desktop/Docker/MessageInput';
+import AudioPlayer from '@/app/home/Docker/AudioPlayer';
+import MessageInput from '@/app/home/Docker/MessageInput';
 import { useConfigStore } from '@/store/config';
 import { useSessionStore } from '@/store/session';
 import { ActionIcon } from '@lobehub/ui';
