+++ src/app/home/VirtualIdol/ImageViewer/style.ts
