+++ src/app/home/VirtualIdol/index.tsx
