+++ src/app/home/VirtualIdol/style.ts
