+++ src/app/page.tsx
@@ -1,5 +1,5 @@
-import PageLoading from './home/loading';
-import Redirect from './home/Redirect';
+import PageLoading from '@/app/welcome/loading';
+import Redirect from '@/app/welcome/Redirect';
 
 const Index = () => (
   <>
