+++ src/panels/ChatPanel/ChatBot/ChatInput/Header/ActionBar/index.tsx
@@ -1,13 +1,11 @@
 import History from './History';
 import Record from './Record';
 import Token from './Token';
-import Voice from './Voice';
 
 const ActionBar = () => (
   <>
     <History />
     <Record />
-    <Voice />
     <Token />
   </>
 );
